requirements.txt
beautifulsoup4==4.6.0
boto3==1.4.7
botocore==1.7.20
Django==1.10.8
django-bootstrap3==9.0.0
django-debug-toolbar==1.8
django-htmlmin==0.10.0
django-redis-cache==1.7.1
django-storages==1.6.5
docutils==0.14
futures==3.1.1
gunicorn==19.7.1
hiredis==0.2.0
html5lib==0.999999999
jmespath==0.9.3
psycopg2==2.7.3
python-dateutil==2.6.1
python-decouple==3.1
pytz==2017.2
redis==2.10.6
s3transfer==0.1.11
six==1.11.0
sqlparse==0.2.3
webencodings==0.5.1
whitenoise==3.3.1
